+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 <!-- next-header -->
 ## [Unreleased]
 
+## [0.4.0] - 2023-05-04
+
 ### Changed
 
 - ⚠️ Changed the meaning of `k` in `smooth(k, n)`.
@@ -58,7 +60,8 @@ and this project adheres to [Semantic Versioning](https://semver.org).
 - Initial implementation of `pratt` and `smooth`.
 
 <!-- next-url -->
-[Unreleased]: https://github.com/FedericoStra/smooth-numbers/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/FedericoStra/smooth-numbers/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/FedericoStra/smooth-numbers/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/FedericoStra/smooth-numbers/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/FedericoStra/smooth-numbers/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/FedericoStra/smooth-numbers/compare/v0.2.0...v0.2.1